Move plugin config related code into new conda.plugins.config module.
